Move to wallet api package
