+++ core/src/main/scala/org/bitcoins/core/api/wallet/db/AddressTagDb.scala
@@ -1,6 +1,12 @@
-package org.bitcoins.core.wallet.utxo
+package org.bitcoins.core.api.wallet.db
 
 import org.bitcoins.core.protocol.BitcoinAddress
+import org.bitcoins.core.wallet.utxo.{
+  AddressTag,
+  AddressTagName,
+  AddressTagType,
+  InternalAddressTag
+}
 
 case class AddressTagDb(
     address: BitcoinAddress,
