+++ wallet-test/src/test/scala/org/bitcoins/wallet/models/AddressTagDAOTest.scala
@@ -2,10 +2,10 @@ package org.bitcoins.wallet.models
 
 import java.sql.SQLException
 
+import org.bitcoins.core.api.wallet.db.AddressTagDb
 import org.bitcoins.core.wallet.utxo.StorageLocationTag.HotStorage
 import org.bitcoins.core.wallet.utxo.{
   AddressTag,
-  AddressTagDb,
   UnknownAddressTag,
   UnknownAddressTagName,
   UnknownAddressTagType
